bug fix v2
 a lot of bug fixing + Logs + Quests + Rank

+++ Villages/src/main/java/pl/kiosel/villages/data/user/top/UserComparator.java
@@ -24,7 +24,7 @@ public final class UserComparator implements TopComparator<UserRank> {
 
     @Override
     public int compare(UserRank o1, UserRank o2) {
-        int result = Float.compare(this.getValue(o1).floatValue(), this.getValue(o2).floatValue());
+		int result = Double.compare(this.getValue(o1).doubleValue(), this.getValue(o2).doubleValue());
         if (result == 0) {
             result = Rank.compareName(o1, o2);
         }
